Remove unnecessary Long→Int conversion in chart data mapping
Vico's series() accepts Number, so Long values can be passed
directly without risking silent overflow for counts above 2.1B.

+++ WordPress/src/main/java/org/wordpress/android/ui/newstats/subscribers/subscribersgraph/SubscribersGraphCard.kt
@@ -237,7 +237,7 @@ private fun SubscribersChart(
         modelProducer.runTransaction {
             lineSeries {
                 series(
-                    dataPoints.map { it.count.toInt() }
+                    dataPoints.map { it.count }
                 )
             }
         }
